add a --dry option so I dont need a github key to run the fuzzer
