+++ reduce_sql.py
@@ -2,7 +2,6 @@ import re
 import subprocess
 import time
 import os
-import fuzzer_helper
 import multiprocessing
 import sqlite3
